gp4par: Fixed signed/unsigned mismatch

+++ src/gp4par/par_main.cpp
@@ -331,7 +331,7 @@ bool PostPARDRC(PARGraph* netlist, Greenpak4Device* device)
 
 	//Check for multiple oscillators with power-down enabled but not the same source
 	powerdowns.clear();
-	for(int i=0; i<device->GetDcmpCount(); i++)
+	for(size_t i=0; i<device->GetDcmpCount(); i++)
 	{
 		auto d = device->GetDcmp(i);
 		if(d->IsUsed())
